server/requirements.txt
Flask[async]
python-kasa
python-dotenv
